Fix transformation of `import * as x from ...` to behave correctly
`import * as x from 'my-module'` imports entire module's contents into
variable `x`. This is represented as ImportDeclaration with
ImportNamespaceSpecifier in Babel AST.

In the current state, this import is transformed in the same way as a
default import (`import x from 'my-module'`) and import of a single export
(`import { x } from 'my-module'`). However, this is incorrect because in
ES modules `import * as x 'my-module'` creates an object with all the
exports of 'my-module'. If 'my-module' contains only the default
export, variable `x` should contain an object with a single property:
`default`.

A raw file is not an ES module, but since we're using `import` syntax
for including it into a module, it should mimics behaviour of importing
an ES module with the default export. That's also how webpack's
raw-loader behaves.

Of course it doesn't make much sense to use this type of import for
"importing" raw files, unless you use something like "import-all.macro"
for importing multiple files that match a glob pattern. And at the same
time you need the import to result in the same thing as when using
webpack's raw-loader or jest-raw-loader.